requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Django==4.1.1
fonttools==4.37.4
google-auth==2.12.0
google-auth-oauthlib==0.5.3
greenlet==1.1.3
gspread==5.5.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
